Add files via upload
Studied the data set, cleaned and prepared the data for visualization.
Found the problem statement that which team of the customer service is not working effectively and which team is quick to respond to the customer. 
Draw my conclusion on the basis of team performance and customer rating. 
Skills used : Analytics, identifying  the problem and understanding the requirement of the company. 
Suggesting points what can increase the growth on the company.